:ok_hand: Update code for better readability
